fix(vercel): add root package.json and move schemas to /src for serverless; update imports; add root tsconfig
