+++ package.json
@@ -1,21 +1,17 @@
 {
-  "name": "prompt2story-vercel",
-  "version": "1.0.0",
-  "description": "Prompt2Story - Vercel serverless functions",
+  "name": "prompt2story",
+  "private": true,
   "type": "module",
-  "engines": {
-    "node": ">=18.0.0"
-  },
   "scripts": {
-    "dev": "vercel dev",
-    "build": "tsc --build"
+    "build": "npm --prefix frontend ci && npm --prefix frontend run build",
+    "typecheck": "npm --prefix frontend run typecheck",
+    "test": "npm --prefix frontend test -- --run"
   },
   "dependencies": {
-    "openai": "^4.67.3",
-    "@vercel/node": "^3.2.20"
+    "openai": "^4.55.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
-    "typescript": "^5.6.2",
-    "@types/node": "^22.10.2"
+    "typescript": "^5.6.2"
   }
 }
